connected the /home page to golang backend. --> *todo: connect form input.
