+++ views/home.html
@@ -1,0 +1,19 @@
+<!DOCTYPE html>
+<html>
+
+<div>
+    <header>Home</header>
+</div>
+
+<div class="form">
+<form action="POST" action="/findStreamer">
+    <input type="text" value="enter streamer username here: frslushh" />
+    <button type="submit" action="submit">submit</button>
+
+</form>
+
+</div>
+
+
+
+</html>
